move model selector to a1/. [skip actions]
